Combine the Iceberg and AWS SDK bridges into one crate
