+++ Cargo.toml
@@ -5,19 +5,18 @@ members = [
     "crates/app",
     "crates/arrow_sql_gen",
     "crates/arrow_tools",
+    "crates/aws-sdk-bridge",
     "crates/data_components",
     "crates/dataformat-json",
     "crates/document_parse",
     "crates/event_stream",
     "crates/flight_client",
     "crates/flightrepl",
-    "crates/iceberg-aws-sdk",
     "crates/llms",
     "crates/model_components",
     "crates/ns_lookup",
     "crates/otel-arrow",
     "crates/runtime-object-store",
-    "crates/object-store-aws-sdk",
     "crates/package",
     "crates/runtime",
     "crates/runtime-auth",
