--- crates/iceberg-aws-sdk/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-edition.workspace = true
-exclude.workspace = true
-homepage.workspace = true
-license.workspace = true
-name = "iceberg-aws-sdk"
-repository.workspace = true
-rust-version.workspace = true
-version.workspace = true
-
-[dependencies]
-anyhow = "1"
-async-trait.workspace = true
-aws-config.workspace = true
-aws-credential-types.workspace = true
-aws-runtime.workspace = true
-aws-sdk-s3.workspace = true
-aws-smithy-runtime.workspace = true
-aws-smithy-runtime-api.workspace = true
-iceberg.workspace = true
-reqwest.workspace = true
-snafu.workspace = true
-tokio.workspace = true
-url.workspace = true
-
-[dev-dependencies]
-aws-sdk-cognitoidentity.workspace = true
-aws-sdk-cognitoidentityprovider.workspace = true
-tempfile.workspace = true
