--- crates/iceberg-aws-sdk/src/lib.rs
@@ -1,41 +1,0 @@
-/*
-Copyright 2025 The Spice.ai OSS Authors
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-     https://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-mod credential_provider;
-use aws_smithy_runtime_api::client::runtime_components::BuildError;
-pub use credential_provider::S3CredentialProvider;
-
-#[derive(Debug, snafu::Snafu)]
-pub enum Error {
-    #[snafu(display(
-        "An unexpected error occurred when initializing the AWS SDK for retrieval of AWS credentials for an Iceberg S3 dataset: {source}."
-    ))]
-    FailedToBuildAWSRuntimeComponents { source: BuildError },
-
-    #[snafu(display(
-        "Failed to find valid credentials from the AWS credential provider chain for the Iceberg S3 connection. Ensure that valid AWS credentials are provided in the environment. Details: https://docs.aws.amazon.com/sdk-for-rust/latest/dg/credproviders.html#credproviders-default-credentials-provider-chain"
-    ))]
-    FailedToGetCredentialsProviderFromConfig,
-
-    #[snafu(display(
-        "Failed to find valid credentials from the AWS credential provider chain for the Iceberg S3 connection. {source} Ensure that valid AWS credentials are provided in the environment. Details: https://docs.aws.amazon.com/sdk-for-rust/latest/dg/credproviders.html#credproviders-default-credentials-provider-chain"
-    ))]
-    FailedToResolveCredentials {
-        source: Box<dyn std::error::Error + Send + Sync>,
-    },
-}
-
-pub type Result<T, E = Error> = std::result::Result<T, E>;
