+++ crates/runtime-object-store/Cargo.toml
@@ -13,6 +13,7 @@ async-stream.workspace = true
 async-trait.workspace = true
 aws-config.workspace = true
 aws-credential-types.workspace = true
+aws-sdk-bridge = { path = "../aws-sdk-bridge" }
 bytes.workspace = true
 chrono = { version = "0.4.38" }
 datafusion.workspace = true
@@ -20,7 +21,6 @@ fundu.workspace = true
 
 futures.workspace = true
 http.workspace = true
-object-store-aws-sdk = { path = "../object-store-aws-sdk" }
 object_store.workspace = true
 reqwest.workspace = true
 serde.workspace = true
