+++ crates/runtime/Cargo.toml
@@ -25,6 +25,7 @@ async-trait.workspace = true
 aws-config = { workspace = true, optional = true }
 aws-credential-types.workspace = true
 aws-sdk-bedrockruntime = { version = "1.55.0", optional = true }
+aws-sdk-bridge = { path = "../aws-sdk-bridge" }
 aws-sdk-dynamodb = { version = "1", optional = true }
 aws-sdk-glue = "1"
 aws-sdk-secretsmanager = { version = "1", optional = true }
@@ -70,7 +71,6 @@ http-body-util = "0.1.3"
 hyper = "1.6.0"
 hyper-util = { version = "0.1.15", features = ["service"] }
 iceberg.workspace = true
-iceberg-aws-sdk = { path = "../iceberg-aws-sdk" }
 iceberg-catalog-glue.workspace = true
 iceberg-catalog-rest.workspace = true
 iceberg-datafusion.workspace = true
@@ -91,7 +91,6 @@ moka.workspace = true
 mysql_async = { workspace = true, optional = true }
 notify = "8.0.0"
 ns_lookup = { path = "../ns_lookup" }
-object-store-aws-sdk = { path = "../object-store-aws-sdk" }
 object_store = { workspace = true, features = ["aws", "http", "azure"] }
 once_cell = "1.21.3"
 opentelemetry.workspace = true
